chunk.h: * Fixed: Chunk.nregs should be u32 in the compiler so overflow doesnt occur.
compiler.h:
* Fixed: CompilerScope.nlocals should also be u32 to prevent oveflow.

inst.h:
* Added: InstGetBX, InstGetAX, InstPatchBX, and InstPatchAX.
* The getters are for decoding larger than 1 byte instruction operands. The patch functions are for upcoming jumps.

+++ src/chunk.h
@@ -5,7 +5,7 @@ typedef u16 ChunkIdx;
 typedef struct Chunk { /* instruction span */
 	InstIdx start;
 	InstIdx len;
-	u8 nregs;
+	u32 nregs;
 } Chunk;
 
 typedef struct Chunks {
